全体のコメント調整 & DescriptorHeap作成関数を設定構造体のみで動くように変更
動作確認済み

+++ RendererInterface/DirectX/DirectXobject/VertexBuffer.cpp
@@ -4,15 +4,22 @@
 
 using namespace render::dx12::object;
 
-///====================================================================
-/// 初期化処理
-///====================================================================
+/* ==================================================================== */
+// Publicメンバー関数
+/* ==================================================================== */
 
-//@brief	=== 設定構造体作成仮想関数 ===
-//@param	data	初期設定データ構造体に設定する先頭ポインター
-//@param	size	初期設定データ構造体に設定するメモリサイズ
-//@return	作成した構造体
-[[nodiscard]] render::dx12::desc::StaticBufferCreateDesc VertexBuffer::create_static_buffer_desc(const void* data, UINT size) {
+/* -- 作成関数 -- */
+
+/// <summary>
+/// 設定構造体作成関数
+/// </summary>
+/// <param name="data">初期設定データ構造体に設定する先頭ポインター</param>
+/// <param name="size">初期設定データ構造体に設定するメモリサイズ</param>
+/// <returns>作成した構造体</returns>
+[[nodiscard]] render::dx12::desc::StaticBufferCreateDesc VertexBuffer::create_static_buffer_desc(
+	const void* data, 
+	UINT64 size
+) {
 
 	//	VerTexBuffer用リソース設定作成
 	render::dx12::desc::StaticBufferCreateDesc desc{};
@@ -29,22 +36,25 @@ using namespace render::dx12::object;
 	return desc;
 }
 
-//@brief	=== 派生先別リソース作成仮想関数 ===
-//@details	基底クラスではS_OKを返す
-//@return	作成の成否
+/// <summary>
+/// 派生先別リソース作成仮想関数
+/// </summary>
+/// <details>
+/// ここでは頂点バッファビューを作成
+/// </details>
+/// <returns>作成の成否</returns>
 [[nodiscard]] HRESULT VertexBuffer::create_resource_object() {
 
 	//	VertexBufferViweを作成
 	vertex_buffer_view.BufferLocation = get_GPU_address();
 	vertex_buffer_view.StrideInBytes = class_size;
-	vertex_buffer_view.SizeInBytes = buffer_size;
+	vertex_buffer_view.SizeInBytes = static_cast<UINT>(buffer_size);
 
 	return S_OK;
 }
 
-///====================================================================
-/// 実行時処理
-///====================================================================
+
+/* -- 取得関数 -- */
 
 //@brief	=== 頂点バッファビュー取得関数 ===
 //@return	頂点バッファビュー参照
